🔨 [packages] Move class `SinglePackageRepository` to fixtures
